% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_number_selection.R
\name{cluster_number_selection}
\alias{cluster_number_selection}
\title{cluster_number_selection}
\usage{
cluster_number_selection(temp_res_gibbs)
}
\arguments{
\item{temp_res_gibbs}{gibbs clustering regression result}
}
\value{
a list consists:
height_record : length(K_seq)*max(K_seq) matrix each row records the height of hclust result
dist_matrix : list with same length as K_seq whose every element consist p dist matrix indicating the distance between the beta(t) of clusters
hclust_trace : hclust results
}
\description{
cluster_number_selection
}
