% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_fun2021.R
\name{clustering_fun2021}
\alias{clustering_fun2021}
\title{comparison algorithm in 2021}
\usage{
clustering_fun2021(reg_dt, label, K, method = "ls")
}
\arguments{
\item{reg_dt}{N*M principle components scores of the observations}

\item{label}{N*1 vector of initialized label sequence}

\item{K}{number of clusters}

\item{method}{loss function used in the regression}
}
\value{
list(label result,estimated coefficients,label record matrix)
}
\description{
comparison algorithm in 2021
}
