% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterwise_reg_2021.R
\name{clusterwise_reg_2021}
\alias{clusterwise_reg_2021}
\title{clusterwise regression algorithm proposed in 2021}
\usage{
clusterwise_reg_2021(
  X_list,
  Y,
  basis,
  sample_t,
  var_ratio,
  K_seq,
  true_label,
  method = "ls",
  ncomp_u = 10,
  ncomp_l = 2,
  seed = 1
)
}
\arguments{
\item{X_list}{list of length K. each element is a N*T matrix. contains simulated X_1:K observed at T evenly spaced time points}

\item{Y}{N*1 vector of the N response variable of the N observations}

\item{basis}{basis used in the smoothing}

\item{sample_t}{T evenly spaced time points of the observations}

\item{var_ratio}{minimum variance demanded to determine how many components are preserved in the FPCA}

\item{K_seq}{int. vector contains the numbers of clusters}

\item{true_label}{N*1 label vector contains ground truth labels}

\item{method}{loss function used in the regression with candidates c("ls","hl","lad")}

\item{ncomp_u}{maximum number of components demanded in FPCA}

\item{ncomp_l}{minimum number of components demanded in FPCA}

\item{seed}{random seed}
}
\value{
list(assessment of the results,BIC_seq,coefficients_list,labels_trace,fpca,components scores,beta_trace,residuals_trace,residuals_list,beta_fd)
assessment of the results: ARI and other assessment standards.
BIC_seq: sequence of BIC with the same length as K_seq
coefficients_list: list consisting of the coefficients matrix estimated under each number of clusters with the same length as K_seq.
label trace: trace of the label sequence throughout sampling process
fpca: fuinctional principle analysis results of X_i(t_j)
components scores: principle components scores of each observations
beta trace: each label sequence of the label trace generates one coefficient estimation
residual trace: each label sequence of the label trace generates one residuals sequence.
beta fd: estimated coefficient in the form of fd object
}
\description{
clusterwise regression algorithm proposed in 2021
}
