% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eBIC.R
\name{eBIC_fun}
\alias{eBIC_fun}
\title{eBIC loss result to determine the number of clusters}
\usage{
eBIC_fun(
  reg_dt,
  label,
  temp_K_seq,
  method = "ls",
  xi_1 = 1,
  xi_2 = 0.5,
  xi_3 = 1
)
}
\arguments{
\item{reg_dt}{N*M principle components scores of the observations}

\item{label}{N*1 label vector that conveys clustering results}

\item{method}{loss function used in the regression}

\item{K}{number of clusters}
}
\value{
calculated BIC results
}
\description{
eBIC loss result to determine the number of clusters
}
