% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_fun.R
\name{estimation_fun}
\alias{estimation_fun}
\title{coefficient estimation of the clustering regression results}
\usage{
estimation_fun(reg_dt, label, mask_index, method = "ls", temp_K_seq)
}
\arguments{
\item{reg_dt}{N*M principle components scores of the observations}

\item{label}{N*1 label vector that conveys clustering results}

\item{mask_index}{int in 1:N used to mask several specific label (for full condition distribution estimation)}

\item{method}{loss function used in the regression}

\item{K}{number of clusters}
}
\value{
list(parameters_mat, residuals_list):
parameters_mat: (M+1)*K coefficient matrix whose first row indicates the intercept term
residuals_list: residuals of the estimation
}
\description{
coefficient estimation of the clustering regression results
}
