% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_clusterwise_regression.R
\name{gibbs_clusterwise_regression}
\alias{gibbs_clusterwise_regression}
\title{integrated gibbs sampling and estimation algorithm for clustering regression}
\usage{
gibbs_clusterwise_regression(
  X_list,
  Y,
  basis,
  sample_t,
  var_ratio = 0.9,
  K_seq,
  true_label,
  rounds = 50,
  burning_period = 10,
  method = "ls",
  lambda = 0.1,
  ncomp_u = 10,
  ncomp_l = 2,
  seed = 1,
  need_trace = F
)
}
\arguments{
\item{X_list}{list of length K. each element is a N*T matrix. contains simulated X_1:K observed at T evenly spaced time points}

\item{Y}{N*1 vector of the N response variable of the N observations}

\item{basis}{basis used in the smoothing}

\item{sample_t}{T evenly spaced time points of the observations}

\item{var_ratio}{minimum variance demanded to determine how many components are preserved in the FPCA}

\item{K_seq}{int. vector contains the numbers of clusters}

\item{true_label}{N*1 label vector contains ground truth labels}

\item{rounds}{R rounds-scalar indicating how many round the algorithm runs}

\item{burning_period}{scalar indicating how many rounds of results are abandoned in the sampling process}

\item{method}{loss function used in the regression with candidates c("ls","hl","lad")}

\item{lambda}{scalar indicating the parameters used to construct a multinomial distribution}

\item{ncomp_u}{maximum number of components demanded in FPCA}

\item{ncomp_l}{minimum number of components demanded in FPCA}

\item{seed}{random seed}
}
\value{
list(assessment of the results,BIC_seq,coefficients_list,labels_trace,fpca,components scores,beta_trace,residuals_trace,residuals_list,beta_fd)
assessment of the results: ARI and other assessment standards.
BIC_seq: sequence of BIC with the same length as K_seq
coefficients_list: list consisting of the coefficients matrix estimated under each number of clusters with the same length as K_seq.
label trace: trace of the label sequence throughout sampling process
fpca: fuinctional principle analysis results of X_i(t_j)
components scores: principle components scores of each observations
beta trace: each label sequence of the label trace generates one coefficient estimation
residual trace: each label sequence of the label trace generates one residuals sequence.
beta fd: estimated coefficient in the form of fd object
}
\description{
integrated gibbs sampling and estimation algorithm for clustering regression
}
