% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_sample.R
\name{gibbs_sample}
\alias{gibbs_sample}
\title{generating R rounds of sample results of labels of the N observations}
\usage{
gibbs_sample(
  reg_dt,
  label_initial,
  lambda = 0.01,
  round_num = 1000,
  seed = 1,
  method = "ls"
)
}
\arguments{
\item{reg_dt}{N*M principle components scores of the observations}

\item{label_initial}{N*1 vector of initialized label sequence}

\item{lambda}{scalar indicating parameters used to construct a multinomial distribution}

\item{round_num}{R rounds-scalar indicating how many round the algorithm runs}

\item{seed}{random seed}

\item{method}{loss function used in the regression}
}
\value{
N*R label matrix indicating the sampling results of each round
}
\description{
generating R rounds of sample results of labels of the N observations
}
