% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_reg_fun.R
\name{hl_reg_fun}
\alias{hl_reg_fun}
\title{regression using Huber loss function(HL)}
\usage{
hl_reg_fun(reg_dt, p)
}
\arguments{
\item{reg_dt}{N*M matrix indicating the M principle scores of N functional samples}
}
\value{
(M+1)*1 vector indicating the regression coefficient whose first elements indicates the intercept term
}
\description{
regression using Huber loss function(HL)
}
