% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_reg_loss.R
\name{hl_reg_loss}
\alias{hl_reg_loss}
\title{residuals of the regression using huber loss function (HL)}
\usage{
hl_reg_loss(reg_dt, p)
}
\arguments{
\item{reg_dt}{N*M matrix indicating the M principle scores of N functional samples}
}
\value{
N*1 vector of residuals of the regression
}
\description{
residuals of the regression using huber loss function (HL)
}
