% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huber_loss.R
\name{huber_loss}
\alias{huber_loss}
\title{Huber loss function to transform residuals}
\usage{
huber_loss(x, c = 1.345)
}
\arguments{
\item{x}{residual term of term model}

\item{c}{huber parameter}
}
\value{
transformed values
}
\description{
Huber loss function to transform residuals
}
