% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_reg_fun.R
\name{logit_reg_fun}
\alias{logit_reg_fun}
\title{regression using least square error(LSE)}
\usage{
logit_reg_fun(reg_dt, p)
}
\arguments{
\item{reg_dt}{N*M matrix indicating the M principle scores of N functional samples}
}
\value{
(M+1)*1 vector indicating the regression coefficient whose first elements indicates the intercept term
}
\description{
regression using least square error(LSE)
}
