% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_reg_loss.R
\name{ls_reg_loss}
\alias{ls_reg_loss}
\title{residuals of the regression using least square error (LS)}
\usage{
ls_reg_loss(reg_dt, p)
}
\arguments{
\item{reg_dt}{N*M matrix indicating the M principle scores of N functional samples}
}
\value{
N*1 vector of residuals of the regression
}
\description{
residuals of the regression using least square error (LS)
}
