% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_res.R
\name{plot_res}
\alias{plot_res}
\title{plot function for gibbs algorithm and its comparision}
\usage{
plot_res(
  function_list,
  temp_res_gibbs,
  temp_res_2021,
  K_seq,
  real_label,
  real_coef_mat,
  y1 = -2,
  y2 = 2,
  y4 = c(-10, 10),
  y5 = c(-10, 10),
  x6 = 5,
  y6 = 5,
  plot_index = c(TRUE, TRUE, TRUE, TRUE, TRUE, FALSE)
)
}
\arguments{
\item{temp_res_gibbs}{result of gibbs algorithm}

\item{temp_res_2021}{result of 2021-comparison algorithm}

\item{K_seq}{int. vector contains the numbers of clusters}

\item{real_label}{N*1 label vector contains ground truth labels}

\item{real_coef_mat}{real coefficient matrix}

\item{y1}{ylim for the 1st plot}

\item{y2}{ylim for the 2nd plot}

\item{x6}{symmetric xlim for the 6th plot}

\item{y6}{ylim for the 6th plot}

\item{plot_index}{logiacal sequence indicating which plot to plot}
}
\value{
automatically generates the plots demanded in the plot_index
}
\description{
plot function for gibbs algorithm and its comparision
}
