% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xij_fun.R
\name{xij_fun}
\alias{xij_fun}
\title{generating simulated X_{ij}(i=1,2,...,N)}
\usage{
xij_fun(function_list, fda_coef_df, sample_t, error_sd, seed = 1)
}
\arguments{
\item{function_list}{K function objects used in simulation compiled into a list}

\item{fda_coef_df}{coefficient matrix with K columns and N rows}

\item{sample_t}{T*1 vector indicating observing time points}

\item{error_sd}{scalar indicating variance of normal observing error}

\item{seed}{random seed}
}
\value{
simulated N*T matrix of sample X_i at time t_j
}
\description{
generating simulated X_{ij}(i=1,2,...,N)
}
