% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yi_fun.R
\name{yi_fun}
\alias{yi_fun}
\title{generating simulated Y_i}
\usage{
yi_fun(
  function_list,
  fda_coef_df_list,
  reg_coef_df_list,
  reg_intercept,
  reg_var,
  reg_error,
  interval,
  seed = 1
)
}
\arguments{
\item{function_list}{K function objected aligned in a list}

\item{fda_coef_df_list}{coefficient matrix with K columns and N rows}

\item{reg_coef_df_list}{N*K coefficient matrix(true coefficient)}

\item{reg_intercept}{scalar indicating the intercept of the regression}

\item{reg_var}{N*1 vector indicating variance parameter of random error term}

\item{reg_error}{N*1 vector indicating random error}

\item{seed}{}
}
\value{
N*1 response variable Y with error term
}
\description{
generating simulated Y_i
}
